improved posts controller

+++ src/controllers/posts.controller.js
@@ -4,10 +4,12 @@ const {
 	models: { User, Post, Favorites, Comments },
 } = require('../sequelize');
 
+const { parseModel, notOwner } = require('../helpers/sequelize');
+
 module.exports = {
 	async getPosts(req, res, next) {
 		try {
-			const posts = await Post.findAll({
+			let posts = await Post.findAll({
 				include: [
 					{
 						model: Favorites,
@@ -19,6 +21,22 @@ module.exports = {
 					},
 				],
 			});
+
+			if (!posts || posts.length === 0) {
+				return res
+					.status(404)
+					.json({ error: true, message: 'No posts found.' });
+			}
+
+			posts = posts.map((post) => parseModel(post));
+			posts = posts.map((post) => ({
+				...post,
+				favorites: post.favorites
+					.map((fav) => parseModel(fav))
+					.map((fav) => fav.userId),
+				comments: post.comments.map((comment) => parseModel(comment)),
+			}));
+
 			res.status(200).json(posts);
 		} catch (e) {
 			next(e);
@@ -27,7 +45,12 @@ module.exports = {
 	async getPost(req, res, next) {
 		const { id } = req.params;
 		try {
-			const post = await Post.findByPk(id, {
+			if (!id || id === '') {
+				return res
+					.status(400)
+					.json({ error: true, message: 'ID of post must be provided' });
+			}
+			let post = await Post.findByPk(id, {
 				include: [
 					{
 						model: Favorites,
@@ -39,15 +62,36 @@ module.exports = {
 					},
 				],
 			});
+			if (!post) {
+				return res.status(404).json({ error: true, message: 'No post found.' });
+			}
+
+			post = parseModel(post);
+			post = {
+				...post,
+				favorites: post.favorites
+					.map((fav) => parseModel(fav))
+					.map((fav) => fav.userId),
+				comments: post.comments.map((comment) => parseModel(comment)),
+			};
 			res.status(200).json(post);
 		} catch (e) {
 			next(e);
 		}
 	},
 	async createPost(req, res, next) {
-		const { title, tags, content } = req.body;
+		let { title, tags, content } = req.body;
 		const { id } = req.user;
 		try {
+			if (!tags || !tags.length) {
+				tags = [];
+			}
+			if (!title || !content) {
+				return res
+					.status(400)
+					.json({ error: true, message: 'All fields must be filled' });
+			}
+
 			const post = await Post.create({
 				title,
 				tags,
@@ -55,6 +99,7 @@ module.exports = {
 				userId: id,
 			});
 			const user = await User.findByPk(id);
+
 			await user.addPost(post);
 			res.status(200).json(post);
 		} catch (e) {
@@ -64,23 +109,54 @@ module.exports = {
 	async updatePost(req, res, next) {
 		let { title, tags, content } = req.body;
 		const { id } = req.params;
+		const userId = req.user.id;
 		try {
-			const postModified = await Post.update(
+			if (!id || id === '') {
+				return res
+					.status(400)
+					.json({ error: true, message: 'ID of post must be provided' });
+			}
+			const found = await Post.findByPk(id);
+			if (!found) {
+				return res.status(404).json({ error: true, message: 'Post not found' });
+			}
+			await Post.update(
 				{ title, tags, content },
 				{ where: { id: { [Op.eq]: id } } }
 			);
-			res.status(200).json(postModified);
+			if (notOwner(userId, found)) {
+				return res
+					.status(401)
+					.json({ error: true, message: 'You are not the owner of this post' });
+			}
+			res.status(200).json({ message: 'Post updated succesfully!' });
 		} catch (e) {
 			next(e);
 		}
 	},
 	async deletePost(req, res, next) {
 		const { id } = req.params;
+		const userId = req.user.id;
 		try {
-			const postModified = await Post.destroy({
+			if (!id || id === '') {
+				return res
+					.status(400)
+					.json({ error: true, message: 'ID of post must be provided' });
+			}
+			const found = await Post.findByPk(id);
+			if (notOwner(userId, found)) {
+				return res
+					.status(401)
+					.json({ error: true, message: 'You are not the owner of this post' });
+			}
+			if (!found) {
+				return res.status(404).json({ error: true, message: 'Post not found' });
+			}
+			await Post.destroy({
 				where: { id: { [Op.eq]: id } },
 			});
-			res.status(204).json(postModified);
+
+			res.status(200).json({ message: 'Post deleted succesfully!' });
 		} catch (e) {
 			next(e);
 		}
